file menu changes - merge the 'upload network from cx' menu item and the 'upload network from table' menu items together - user can click  'import network from file' - choosing a 'cx2' file performs the import cx2 function - choosing a tabular data file starts the table data loader form
