+++ src/components/ToolBar/DataMenu/LoadFromNdexMenuItem.tsx
@@ -17,7 +17,7 @@ export const LoadFromNdexMenuItem = (props: BaseMenuProps): ReactElement => {
 
   return (
     <>
-      <MenuItem onClick={handleOpenDialog}>Network from NDEx</MenuItem>
+      <MenuItem onClick={handleOpenDialog}>Open Network(s) from NDEx</MenuItem>
       <LoadFromNdexDialog open={openDialog} handleClose={handleCloseDialog} />
     </>
   )
